requirements.txt
# Dependencies
streamlit==1.29.0
openai==1.3.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0